1. Added Compilation controller and service 2. Added TmdbService: api to TMDB 3. Added errors for the Compilation controller and TmdbService 4. Fixed some bugs related to auth 5. Fixed some DTO related to Compilation 6. Added some repositories 7. Added ModelMapper and RestTemplate

+++ src/main/java/app/viewflowbackend/DTO/auxiliary/CompilationMediaDTO.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CompilationMediaDTO {
-    private Integer mediaId;
+    private Long mediaId;
 
     private MediaType mediaType;
